front, back 연동

+++ javaStudy/React/spring_boot_store_tcode/src/List.js
@@ -6,7 +6,27 @@ export default function List(){
     const dispatch = useDispatch();
 
     const handleDelete = (e)=>{
-        dispatch(productDelete(e.target.id));
+		fetch("http://localhost:8080/product/" + e.target.id, {
+			method: "DELETE",
+		})
+		.then(response => {
+			if (!response.ok) {
+				throw new Error("네트워크 오류");
+			}
+			return response.text(); // String으로 보냈기 때문에 json아니고 text
+		})
+		.then(data => {
+			// 그냥 dispatch만 쓰면 상품 삭제 성공했을 때 실패했을 때 두 가지 경우 다 들어옴. => if문으로 먼저 확인하기
+			if (data === "Product deleted successfully") {
+				dispatch(productDelete(e.target.id));
+				console.log(data);
+			} else {
+				console.log(data);
+			}
+		})
+		.catch(error => {
+			console.log(error);
+		});
     };
 
     const list = pdList.map(t=>(
